# springDemo/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8030",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/student": {
      "get": {
        "tags": [
          "student-controller"
        ],
        "operationId": "getStudent",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "student-controller"
        ],
        "operationId": "postStudent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "StudentModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "subject": {
            "type": "string"
          }
        }
      }
    }
  }
}